# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "question": {
            "default": "",
            "title": "Question",
            "type": "string"
          },
          "repository_id": {
            "format": "uuid",
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "repository_id",
          "conversation_id"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "Conversation": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "A UUID version 4 identifier.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "repository_id": {
            "description": "A UUID version 4 identifier.",
            "format": "uuid",
            "title": "Repository Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repository_id",
          "created_at",
          "updated_at"
        ],
        "title": "Conversation",
        "type": "object"
      },
      "ConversationRequest": {
        "additionalProperties": false,
        "properties": {
          "repository_id": {
            "format": "uuid",
            "title": "Repository Id",
            "type": "string"
          }
        },
        "required": [
          "repository_id"
        ],
        "title": "ConversationRequest",
        "type": "object"
      },
      "FileResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "FileResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndexingError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retriable": {
            "default": false,
            "title": "Retriable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "retriable"
        ],
        "title": "IndexingError",
        "type": "object"
      },
      "IndexingJob": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IndexingError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "description": "A UUID version 4 identifier.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "repository_id": {
            "description": "A UUID version 4 identifier.",
            "format": "uuid",
            "title": "Repository Id",
            "type": "string"
          },
          "stage": {
            "enum": [
              "queued",
              "cloning",
              "discovering",
              "parsing",
              "indexing",
              "finalizing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "indexing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repository_id",
          "status",
          "stage",
          "progress",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at",
          "error"
        ],
        "title": "IndexingJob",
        "type": "object"
      },
...
      "Repository": {
        "additionalProperties": false,
        "properties": {
          "default_branch": {
            "minLength": 1,
            "title": "Default Branch",
            "type": "string"
          },
          "id": {
            "description": "A UUID version 4 identifier.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "indexed_revision": {
            "minLength": 1,
            "title": "Indexed Revision",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "minLength": 1,
            "title": "Owner",
            "type": "string"
          },
          "source_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_url",
          "owner",
          "name",
          "default_branch",
          "indexed_revision"
        ],
        "title": "Repository",
        "type": "object"
      },
      "RepositoryCreateResponse": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "job": {
            "$ref": "#/components/schemas/IndexingJob"
          },
          "repository": {
            "$ref": "#/components/schemas/Repository"
          }
        },
        "required": [
          "repository",
          "conversation",
          "job"
        ],
        "title": "RepositoryCreateResponse",
        "type": "object"
      },
      "RepositoryRequest": {
        "additionalProperties": false,
        "properties": {
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          }
        },
        "title": "RepositoryRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceResponse": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "job": {
            "$ref": "#/components/schemas/IndexingJob"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Messages",
            "type": "array"
          },
          "repository": {
            "$ref": "#/components/schemas/Repository"
          },
          "selected_file": {
            "title": "Selected File",
            "type": "string"
          },
          "starter_questions": {
            "items": {
              "type": "string"
            },
            "title": "Starter Questions",
            "type": "array"
          },
          "tree": {
            "items": {
              "$ref": "#/components/schemas/TreeNode"
            },
            "title": "Tree",
            "type": "array"
          }
        },
        "required": [
          "repository",
          "conversation",
          "tree",
          "selected_file",
          "starter_questions",
          "messages",
          "job"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Catch-up backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chat/stream": {
      "post": {
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream"
      }
    },
    "/api/conversations": {
      "post": {
        "operationId": "create_conversation_api_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job"
      }
    },
    "/api/repositories": {
      "post": {
        "operationId": "create_repository_api_repositories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Repository"
      }
    },
    "/api/repositories/{owner}/{repo}/files": {
      "get": {
        "operationId": "get_file_api_repositories__owner___repo__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get File"
      }
    },
    "/api/repositories/{owner}/{repo}/workspace": {
      "get": {
        "operationId": "get_workspace_api_repositories__owner___repo__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "repo",
            "required": true,
            "schema": {
              "title": "Repo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace"
      }
    },
    "/api/repositories/{repository_id}/indexing-jobs": {
      "post": {
        "operationId": "create_indexing_job_api_repositories__repository_id__indexing_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJob"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Indexing Job"
      }
    },
...
  }
}